Allow data holes
